requirements.txt
setuptools
uvicorn
pydantic
fastapi
pyro-ppl
torch
nlprep
